+++ src/common/monnef/jaffas/trees/BlockFruitSapling.java
@@ -14,7 +14,7 @@ public class BlockFruitSapling extends BlockFlower {
     public boolean onBlockActivated(World par1World, int par2, int par3, int par4, EntityPlayer par5EntityPlayer, int par6, float par7, float par8, float par9) {
         // bonemeal
         ItemStack itemstack = par5EntityPlayer.inventory.getCurrentItem();
-        if (itemstack != null && itemstack.itemID == Item.dyePowder.shiftedIndex) {
+        if (itemstack != null && itemstack.itemID == Item.dyePowder.shiftedIndex && mod_jaffas_trees.bonemealingAllowed) {
             if (itemstack.getItemDamage() == 15) {
                 growTree(par1World, par2, par3, par4, rand);
                 itemstack.stackSize--;
@@ -115,9 +115,9 @@ public class BlockFruitSapling extends BlockFlower {
      * returns a list of blocks with the same ID, but different meta (eg: wood returns 4 blocks)
      */
     public void getSubBlocks(int par1, CreativeTabs par2CreativeTabs, List par3List) {
-        par3List.add(new ItemStack(par1, 1, 0));
+/*        par3List.add(new ItemStack(par1, 1, 0));
         par3List.add(new ItemStack(par1, 1, 1));
         par3List.add(new ItemStack(par1, 1, 2));
-        par3List.add(new ItemStack(par1, 1, 3));
+        par3List.add(new ItemStack(par1, 1, 3));*/
     }
 }